Fix avoid_quick_create with multiprocess
When starting Odoo in multiprocess, a change to the field
IrModel.avoid_quick_create would be active only on the current worker.
Otherwise, it would be active after a restart.

The proposed change is to signal a registry change, forcing the other
workers to reload a registry and then be aware of the change.