[FIX] base_search_custom_field_filter: Avoid problem with records with same field content
If the field you are looking for is a relational one, and contains the same value
in different records, when you unfold the results and select one of them, the results
of the search belongs to the first ocurrence instead the specific selected record.

This is due to Odoo JS code using same operator in this kind of search when filter_domain
is used in the filter. We avoid this letting the standard field code to act, but
injecting a custom attr that overwrites the left leaf of the domain.