Reject Tier Validation
- Fix a bug when a definition's review_type is set to expression, the
user(s) are unable to reject the record.

When rejecting a record, the reviews are filtered using the records
reviewer_id and reviewer_group_id which are only set when the review
type is set to individual or group. To support other review types,
the filter is being changed to check the current user against the
review_ids field instead.