[FIX] date_range: Failed compute in required field
We can't assign an empty recordset in a computed writable for a required field for
triggering DB fault, as previously this we get an akward error no in current
ORM status:

ERROR: operator does not exist: integer = boolean
LINE 12:                     AND dt.type_id=false;
                                           ^
HINT:  No operator matches the given name and argument types. You might need to add explicit type casts.

so let's simply trigger in this case the check of the company consistency, which is
legit and the final expected result.

I have added an extra test for testing the company consistency the other way around.