[FIX] mass_editing: Unsatisfied NOT NULL constraint on action name
Since this module was refactored and it now depends on
`mass_operation_abstract`, a new required field `action_name` is
introduced, that might not be already filled. That causes the SQL
constraint to fail to be applied, showing the following message:

    odoo.modules.loading: Table 'mass_editing': column 'action_name': unable to set constraint NOT NULL
    column "action_name" contains null values

To solve the above, the new field is initialized with the same value as
name, in case it's not filled yet.