[FIX] base_duplicate_security_group: fix test errors
When the web_responsive module is installed, the tests fail when looking
for "Settings" in the menu, because you have to access the menu first
to get that option. Adding the extra step to display the menu is the
right option so that the tests don't fail whether the web_responsive
module is installed or not.

TT42816