[FIX] base_tier_validation: make ReviewsTable widget work with manual config
When an Odoo model inherits from `tier.validation` but sets up
`_tier_validation_manual_config = True`, adding `review_ids` in a form
view with the `tier_validation` widget causes Owl errors.

It appears that the Tier Review fields are then not passed to the Owl
framework, and accessing them in the widget template raises exceptions.

Adding the fields used in the `base_tier_validation.Collapse` template
to the component under the `relatedFields` key fixes this, while not
changing the behavior for models using `_tier_validation_manual_config =
False`.