[FIX] base_tier_validation: prevent error when trying to access review_ids in a one2many
When the validations are in a model that is inside another model, after the reviews are approved, trying to open the one2many raises an error. This commit ensures that the value is taken from the correct record instead of env.model.root, which represents the main model and not the model that inherits tier.validation.